+++ async-race/src/index.html
@@ -9,6 +9,6 @@
 </head>
 <body>
     <div class = "wrapper"></div>
-    <script type="module" src="./js/index.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
